Add another facility page test
